test: replace vacuous mulligan-skip test with a direct filter unit test (#224)
The previous `mulligan_loop_skips_defeated_investigator_in_turn_order`
test was vacuous: it used `.with_mulligan_remaining([inv2])` to seed the
state and then asserted `current_mulligan() == Some(inv2)`, which only
proved the builder set the field — no engine function was called, so the
filter under test (`active_investigators_in_turn_order`) was never exercised.

Replace it with a focused unit test directly in
`dispatch/cursor.rs::tests::active_investigators_in_turn_order_excludes_eliminated`
that calls the real filter function and asserts the Killed investigator
is excluded from the result. The test will fail if the Killed-exclusion
predicate is removed.

Also corrects the leftover comment near the deleted
`non_resolve_input_action_rejects_while_skill_test_paused` site to
accurately cite `crates/game-core/tests/reaction_windows.rs::
non_resolve_input_action_rejects_while_window_open` as the actual
coverage point for the pending-prompt-gate rejection branch.
